feat: 新增班级管理相关功能并优化课程查询逻辑
1.  新增班级与老师关联表相关代码，包括实体类、Mapper、Service及实现
2.  新增班级插入功能，支持批量导入班级排班和关联老师
3.  为课程查询DTO新增搜索关键词字段，优化课程列表查询支持模糊搜索
4.  调整Class实体类字段类型，将学生人数相关字段改为Long类型
5.  修复学生人数统计AOP的类型转换问题
6.  调整MyBatis-Plus依赖导入顺序，优化代码结构
7.  新增班级插入接口和对应的Controller映射

+++ src/main/java/com/shiroko/aspect/StudentCountAspect.java
@@ -17,7 +17,10 @@ import org.springframework.stereotype.Component;
 import java.lang.reflect.Field;
 
 /**
- * Description: TODO
+ * Description: 班级人数更新 AOP
+ * 1. 监听 UpdateStudentCount 注解
+ * 2. 从中间表（class_student）统计人数
+ * 3. 更新到 Class 表（实体类名建议改为 Clazz 避免关键字冲突）
  *
  * @author Guguguy
  * @version 1.0
@@ -50,7 +53,7 @@ public class StudentCountAspect {
 
             // 3. 更新到 Class 表（实体类名建议改为 Clazz 避免关键字冲突）
             classMapper.update(null, new LambdaUpdateWrapper<Class>()
-                    .set(Class::getStudentCount, currentCount.intValue())
+                    .set(Class::getStudentCount, currentCount)
                     .eq(Class::getId, classId));
 
             log.info("班级ID: {} 人数已更新为: {}", classId, currentCount);
